+++ packages/browser/lib/utils/concatUint8Arrays.ts
@@ -1,4 +1,4 @@
-export default function concatUint8Arrays(arrays: Uint8Array[]) {
+export default function concatUint8Arrays(arrays: Uint8Array[]): Uint8Array {
   const totalLength = arrays.reduce((acc, array) => acc + array.length, 0);
   const result = new Uint8Array(totalLength);
   let offset = 0;
