+++ packages/browser/lib/utils/index.ts
@@ -1,7 +1,12 @@
 export * as Datatypes from "./datatypes";
 export * as errors from "./errors";
-export * as cryptoUtils from "./crypto";
+
 
 export { default as extractDomain } from "./extractDomain";
 export { default as calculateHeight } from "./calculateHeight";
 export { default as constructSource } from "./constructSource";
+export { default as concatUint8Arrays } from "./concatUint8Arrays";
+export { default as ecPointCompress } from "./ecPointCompress";
+export { default as buildCageKeyFromSuppliedPublicKey } from "./buildCageKeyFromSuppliedPublicKey";
+export { default as deriveSharedSecret } from "./deriveSharedSecret";
+export { default as crc32 } from "./crc32";
